requirements.txt
Flask==1.0.2
requests==2.21.0
python-twitter==3.5
google-cloud-firestore==0.32.1
